requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
streamlit==1.28.1
pandas==2.1.3
plotly==5.17.0
python-dotenv==1.0.0
groq==0.4.1
google-generativeai==0.3.2
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
tenacity==8.2.3
pymongo==4.6.0
dnspython==2.4.0
